Crate libappindicator_sys

Structs

_AppIndicator

AppIndicator:

_AppIndicatorPrivate

_GData

_GObject

GObject:

_GParamSpec

GParamSpec: (ref-func g_param_spec_ref_sink) (unref-func g_param_spec_uref) (set-value-func g_value_set_param) (get-value-func g_value_get_param) @g_type_instance: private #GTypeInstance portion @name: name of this parameter: always an interned string @flags: #GParamFlags flags for this parameter @value_type: the #GValue type for this parameter @owner_type: #GType type that uses (introduces) this parameter

_GSList

_GTypeClass

GTypeClass:

_GTypeInstance

GTypeInstance:

_GValue

GValue:

_GtkContainer

_GtkContainerPrivate

_GtkMenu

_GtkMenuPrivate

_GtkMenuShell

_GtkMenuShellPrivate

_GtkStatusIcon

_GtkStatusIconPrivate

_GtkWidget

_GtkWidgetPrivate

Constants

AppIndicatorCategory_APP_INDICATOR_CATEGORY_APPLICATION_STATUS

AppIndicatorCategory_APP_INDICATOR_CATEGORY_COMMUNICATIONS

AppIndicatorCategory_APP_INDICATOR_CATEGORY_HARDWARE

AppIndicatorCategory_APP_INDICATOR_CATEGORY_OTHER

AppIndicatorCategory_APP_INDICATOR_CATEGORY_SYSTEM_SERVICES

AppIndicatorStatus_APP_INDICATOR_STATUS_ACTIVE

AppIndicatorStatus_APP_INDICATOR_STATUS_ATTENTION

AppIndicatorStatus_APP_INDICATOR_STATUS_PASSIVE

GParamFlags_G_PARAM_CONSTRUCT

GParamFlags_G_PARAM_CONSTRUCT_ONLY

GParamFlags_G_PARAM_DEPRECATED

GParamFlags_G_PARAM_EXPLICIT_NOTIFY

GParamFlags_G_PARAM_LAX_VALIDATION

GParamFlags_G_PARAM_PRIVATE

GParamFlags_G_PARAM_READABLE

GParamFlags_G_PARAM_READWRITE

GParamFlags_G_PARAM_STATIC_BLURB

GParamFlags_G_PARAM_STATIC_NAME

GParamFlags_G_PARAM_STATIC_NICK

GParamFlags_G_PARAM_WRITABLE

GdkScrollDirection_GDK_SCROLL_DOWN

GdkScrollDirection_GDK_SCROLL_LEFT

GdkScrollDirection_GDK_SCROLL_RIGHT

GdkScrollDirection_GDK_SCROLL_SMOOTH

GdkScrollDirection_GDK_SCROLL_UP

Statics

LIB

Functions

app_indicator_build_menu_from_desktop⚠

app_indicator_get_attention_icon⚠

app_indicator_get_attention_icon_desc⚠

app_indicator_get_category⚠

app_indicator_get_icon⚠

app_indicator_get_icon_desc⚠

app_indicator_get_icon_theme_path⚠

app_indicator_get_id⚠

app_indicator_get_label⚠

app_indicator_get_label_guide⚠

app_indicator_get_menu⚠

app_indicator_get_ordering_index⚠

app_indicator_get_secondary_activate_target⚠

app_indicator_get_status⚠

app_indicator_get_title⚠

app_indicator_get_type⚠

app_indicator_new⚠

app_indicator_new_with_path⚠

app_indicator_set_attention_icon⚠

app_indicator_set_attention_icon_full⚠

app_indicator_set_icon⚠

app_indicator_set_icon_full⚠

app_indicator_set_icon_theme_path⚠

app_indicator_set_label⚠

app_indicator_set_menu⚠

app_indicator_set_ordering_index⚠

app_indicator_set_secondary_activate_target⚠

app_indicator_set_status⚠

app_indicator_set_title⚠

Type Aliases

AppIndicator

AppIndicatorCategory

AppIndicatorCategory: @APP_INDICATOR_CATEGORY_APPLICATION_STATUS: The indicator is used to display the status of the application. @APP_INDICATOR_CATEGORY_COMMUNICATIONS: The application is used for communication with other people. @APP_INDICATOR_CATEGORY_SYSTEM_SERVICES: A system indicator relating to something in the user’s system. @APP_INDICATOR_CATEGORY_HARDWARE: An indicator relating to the user’s hardware. @APP_INDICATOR_CATEGORY_OTHER: Something not defined in this enum, please don’t use unless you really need it.

AppIndicatorPrivate

AppIndicatorStatus

AppIndicatorStatus: @APP_INDICATOR_STATUS_PASSIVE: The indicator should not be shown to the user. @APP_INDICATOR_STATUS_ACTIVE: The indicator should be shown in it’s default state. @APP_INDICATOR_STATUS_ATTENTION: The indicator should show it’s attention icon.

GData

GInitiallyUnowned

GObject

GParamFlags

GParamFlags: @G_PARAM_READABLE: the parameter is readable @G_PARAM_WRITABLE: the parameter is writable @G_PARAM_READWRITE: alias for %G_PARAM_READABLE | %G_PARAM_WRITABLE @G_PARAM_CONSTRUCT: the parameter will be set upon object construction @G_PARAM_CONSTRUCT_ONLY: the parameter can only be set upon object construction @G_PARAM_LAX_VALIDATION: upon parameter conversion (see g_param_value_convert()) strict validation is not required @G_PARAM_STATIC_NAME: the string used as name when constructing the parameter is guaranteed to remain valid and unmodified for the lifetime of the parameter. Since 2.8 @G_PARAM_STATIC_NICK: the string used as nick when constructing the parameter is guaranteed to remain valid and unmmodified for the lifetime of the parameter. Since 2.8 @G_PARAM_STATIC_BLURB: the string used as blurb when constructing the parameter is guaranteed to remain valid and unmodified for the lifetime of the parameter. Since 2.8 @G_PARAM_EXPLICIT_NOTIFY: calls to g_object_set_property() for this property will not automatically result in a “notify” signal being emitted: the implementation must call g_object_notify() themselves in case the property actually changes. Since: 2.42. @G_PARAM_PRIVATE: internal @G_PARAM_DEPRECATED: the parameter is deprecated and will be removed in a future version. A warning will be generated if it is used while running with G_ENABLE_DIAGNOSTIC=1. Since 2.26

GParamSpec

GSList

GType

GTypeClass

GTypeInstance

GValue

GdkScrollDirection

GdkScrollDirection: @GDK_SCROLL_UP: the window is scrolled up. @GDK_SCROLL_DOWN: the window is scrolled down. @GDK_SCROLL_LEFT: the window is scrolled to the left. @GDK_SCROLL_RIGHT: the window is scrolled to the right. @GDK_SCROLL_SMOOTH: the scrolling is determined by the delta values in #GdkEventScroll. See gdk_event_get_scroll_deltas(). Since: 3.4

gboolean

gchar

gdouble

gfloat

gint

gint64

glong

gpointer

gsize

guint

guint32

guint64

gulong

Unions

_GValue__bindgen_ty_1